refact(dim): remove completeDimensions
